Extracted JsonApiReader
